DRY -> Don't Repeat Yourself -> a good thing to always remember
If you need to write the same thing more than twice, you may be able to generalize it and save yourself some trouble.

Here, I've added a new class and use it to style all the colour-guide squares, removing many lines of css, and making it easier to edit the design, if needed.